Moved helper staticmethods to a dedicated module
